+++ examples/random_client.php
@@ -149,7 +149,7 @@ while (true) {
                     break;
                 case 6:
                     echo "< Throw CloseException\n";
-                    throw new CloseException(1008, 'Closing with exception');
+                    throw new CloseException('Closing with exception', status: 1008);
                 case 7:
                     echo "< Throw ReconnectException\n";
                     throw new ReconnectException();