+++ src/Client.php
@@ -470,7 +470,7 @@ class Client implements IdentityInterface, LoggerAwareInterface, Stringable
             ->withScheme(match ($this->socketUri->getScheme()) {
                 'ws', 'http' => 'tcp',
                 'wss', 'https' => 'ssl',
-                default => throw new ClientException("Invalid socket scheme: {$this->socketUri->getScheme()}")
+                default => throw new ClientException('Invalid scheme on {uri}', uri: $this->socketUri)
             })
             ->withHost($this->socketUri->getHost(Uri::IDN_ENCODE))
             ->withPort($this->socketUri->getPort(Uri::REQUIRE_PORT));
@@ -483,14 +483,13 @@ class Client implements IdentityInterface, LoggerAwareInterface, Stringable
             $client->setTimeout($this->configuration->getTimeout());
             $stream = $client->connect();
         } catch (Throwable $e) {
-            $error = "Could not open socket to \"{$hostUri}\": {$e->getMessage()}";
             $this->configuration->getLogger()->error("[{scope}] {message}", [
                 'scope' => self::SCOPE,
                 'client' => $this->identity,
                 'exception' => $e,
                 'message' => $e->getMessage(),
             ]);
-            throw new ClientException($error);
+            throw new ClientException('Could not connect to {uri}', uri: $hostUri);
         }
         $connection = $this->connections->create($stream, $hostUri->getScheme() === 'ssl');
         $this->connections->attach($connection);
@@ -536,7 +535,7 @@ class Client implements IdentityInterface, LoggerAwareInterface, Stringable
                 'connection' => $connection->getIdentity(),
                 'uri' => $hostUri,
             ]);
-            throw new ClientException("Invalid stream on \"{$hostUri}\".");
+            throw new ClientException('Invalid stream on {uri}', uri: $hostUri);
         }
         try {
             if (!$this->configuration->isPersistent() || $stream->tell() == 0) {
@@ -677,13 +676,18 @@ class Client implements IdentityInterface, LoggerAwareInterface, Stringable
             /** @var ResponseInterface */
             $response = $connection->pullHttp();
             if ($response->getStatusCode() != 101) {
-                throw new HandshakeException("Invalid status code {$response->getStatusCode()}.", $response);
+                throw new HandshakeException(
+                    'Invalid status code {statusCode}',
+                    response: $response,
+                    statusCode: $response->getStatusCode(),
+                );
             }
 
             if (empty($response->getHeaderLine('Sec-WebSocket-Accept'))) {
                 throw new HandshakeException(
-                    "Connection to '{$uri}' failed: Server sent invalid upgrade response.",
-                    $response
+                    'Connection to {uri} failed: Invalid upgrade response',
+                    uri: $uri,
+                    response: $response,
                 );
             }
 
@@ -693,7 +697,10 @@ class Client implements IdentityInterface, LoggerAwareInterface, Stringable
             );
 
             if ($responseKey !== $expectedKey) {
-                throw new HandshakeException("Server sent bad upgrade response.", $response);
+                throw new HandshakeException(
+                    'Server sent bad upgrade response',
+                    response: $response,
+                );
             }
         } catch (HandshakeException $e) {
             $this->configuration->getLogger()->error("[{scope}] {message}", [